+++ ingredientCategory.py
@@ -3,14 +3,15 @@ from model import IngredientCategory
 import jinja2
 import os
 from base_handler import BaseHandler
-from user_management import isUserAdmin
+from user_management import isUserAdmin, isUserCook
 
 jinja_environment = jinja2.Environment(loader=jinja2.FileSystemLoader(os.path.dirname(__file__)))
 
 class CategoryIngredientDeletePage(BaseHandler):
 	def post(self):
-		if(not isUserAdmin(self)):
-			self.redirect("/")	
+		if not isUserCook(self):
+			self.redirect("/")
+			return
 		category = db.get(self.request.get('ingredientCategoryKey'))
 		ingredient = db.get(self.request.get('ingredientKey'))
 		ingredient.category=None
@@ -19,15 +20,17 @@ class CategoryIngredientDeletePage(BaseHandler):
 
 class IngredientCategoryPage(BaseHandler):
 	def post(self):
-		if(not isUserAdmin(self)):
-			self.redirect("/")	
+		if not isUserCook(self):
+			self.redirect("/")
+			return
 		ingredientCategory = IngredientCategory()
 		ingredientCategory.name = self.request.get('ingredient_category_name')
 		ingredientCategory.put()
 		self.redirect('/ingredientCategory')
 	def get(self):
-		if(not isUserAdmin(self)):
-			self.redirect("/")	
+		if not isUserCook(self):
+			self.redirect("/")
+			return
 		if ((self.request.get('ingredientCategoryKey') != None) and (self.request.get('ingredientCategoryKey') != "")):
 		# List every ingredient in the category
 			ingredientCategory = db.get(self.request.get('ingredientCategoryKey'))
@@ -50,8 +53,9 @@ class IngredientCategoryPage(BaseHandler):
 
 class IngredientCategoryDeletePage(BaseHandler):
 	def post(self):
-		if(not isUserAdmin(self)):
-			self.redirect("/")	
+		if not isUserCook(self):
+			self.redirect("/")
+			return
 		ingredientCategory = db.get(self.request.get('ingredientCategoryKey'))	  
 		ingredientCategory.delete()
 		self.redirect('/ingredientCategory')
